+++ classes/sdk/Serializers.php
@@ -10,7 +10,8 @@ class Aplazame_Serializers
             require_once(ABSPATH . 'wp-admin/includes/plugin.php');
         }
 
-        return get_plugins('/woocommerce')['woocommerce.php']['Version'];
+        $woo = get_plugins('/woocommerce');
+        return $woo['woocommerce.php']['Version'];
     }
 
     public static function get_meta()
@@ -52,18 +53,24 @@ class Aplazame_Serializers
         return $articles;
     }
 
-    public function get_customer($customer)
+    public function get_user($user)
     {
-        $serializer = array(
-            'id' => (string) $customer->id,
+        return array(
+            'id' => (string) $user->id,
+            'type' => 'e',
+            'gender' => 0,
+            'email' => $user->user_email,
+            'first_name' => $user->first_name,
+            'last_name' => $user->last_name,
+            'date_joined' => date(DATE_ISO8601, $user->user_registered));
+    }
+
+    public function get_customer($billing_email)
+    {
+        return array(
             'type' => 'n',
             'gender' => 0,
-            'email' => $customer->user_email,
-            'first_name' => $customer->first_name,
-            'last_name' => $customer->last_name,
-            'date_joined' => date(DATE_ISO8601, $customer->user_registered));
-
-        return $serializer;
+            'email' => $billing_email);
     }
 
     public function get_address($order, $type)
@@ -106,7 +113,7 @@ class Aplazame_Serializers
     public function get_order($order)
     {
         $serializer = array(
-            'id' => $order->id,
+            'id' => (string) $order->id,
             'articles' => $this->get_articles($order->get_items()),
             'currency' => get_woocommerce_currency(),
             'total_amount' => Aplazame_Filters::decimals($order->get_total()),
@@ -116,9 +123,9 @@ class Aplazame_Serializers
         return $serializer;
     }
 
-    public function get_checkout($order, $checkout_url, $customer)
+    public function get_checkout($order, $checkout_url, $user)
     {
-        return array(
+        $serializer = array(
             'toc' => true,
             'merchant' => array(
                 'confirmation_url' => home_url(
@@ -130,11 +137,19 @@ class Aplazame_Serializers
                 'success_url' => html_entity_decode(
                     $order->get_checkout_order_received_url())
             ),
-            'customer' => $this->get_customer($customer),
+            'customer' => $user->id?$this->get_user(
+                $user):$this->get_customer($order->billing_email),
             'order' => $this->get_order($order),
             'billing' => $this->get_address($order, 'billing'),
-            'shipping' => $this->get_shipping_info($order),
             'meta' => static::get_meta());
+
+        $shipping_method = $order->get_shipping_method();
+
+        if (!empty($shipping_method)) {
+            $serializer['shipping'] = $this->get_shipping_info($order);
+        }
+
+        return $serializer;
     }
 
     public function get_history($qs)